+++ OneBlink.SDK/models/FormElementAttributeMapping.cs
@@ -4,7 +4,7 @@ namespace OneBlink.SDK.Model
 {
     public class FormElementAttributeMapping
     {
-        public Guid elementId { get; set; }
+        public string elementId { get; set; }
         public string @attribute { get; set; }
     }
 }