+++ OneBlink.SDK/models/FormElementOptionAttribute.cs
@@ -6,6 +6,6 @@ namespace OneBlink.SDK.Model
     public class FormElementOptionAttribute
     {
         public List<string> optionIds { get; set; }
-        public Guid elementId { get; set; }
+        public string elementId { get; set; }
     }
 }