+++ OneBlink.SDK/models/FormsAppMenuItem.cs
@@ -18,7 +18,7 @@ namespace OneBlink.SDK.Model
     {
         public string type { get; set; }
         public string attribute { get; set; }
-        public Guid? formElementId { get; set; }
+        public string formElementId { get; set; }
     }
 
     public class FormsAppMenuItem